git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1688 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ include/avisynth/avisynth1.h
@@ -135,7 +135,7 @@ class VideoFrame {
 
   VideoFrame(VideoFrameBuffer* _vfb, int _offset, int _pitch, int _row_size, int _height);
 
-  void* operator new(unsigned size);
+  void* operator new(size_t size);
 
 public:
   int GetPitch() const { return pitch; }
