+++ src/subtitles/GFN.cpp
@@ -85,7 +85,7 @@ void GetSubFileNames(CString fn, CAtlArray<CString>& paths, CAtlArray<SubFile>& 
 		WIN32_FIND_DATA wfd, wfd2;
 		HANDLE hFile, hFile2;
 
-		for(int k = 0; k < paths.GetCount(); k++)
+		for(ptrdiff_t k = 0; k < paths.GetCount(); k++)
 		{
 			CString path = paths[k];
 			path.Replace('\\', '/');
@@ -120,9 +120,9 @@ void GetSubFileNames(CString fn, CAtlArray<CString>& paths, CAtlArray<SubFile>& 
 			// TODO: use 'sl' in the next step to find files (already a nice speedup as it is now...)
 			if(fEmpty) continue;
 
-			for(int j = 0; j < extlistnum; j++)
+			for(ptrdiff_t j = 0; j < extlistnum; j++)
 			{
-				for(int i = 0; i < extsubnum; i++)
+				for(ptrdiff_t i = 0; i < extsubnum; i++)
 				{
 					if((hFile = FindFirstFile(path + title + ext[j][i], &wfd)) != INVALID_HANDLE_VALUE)
 					{
